commit first

+++ src/main/java/ru/kata/spring/boot_security/demo/rep/RoleRepository.java
@@ -3,6 +3,6 @@ package ru.kata.spring.boot_security.demo.rep;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.kata.spring.boot_security.demo.model.Role;
 
-public interface RoleRepository extends JpaRepository<Role, Integer> {
-    Role findByName(String name);
+public interface RoleRepository extends JpaRepository<Role, Long> {
+  Role findByName(String name);
 }
